Moved Item.ts to types
